# api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/validateotp:
    post:
      tags:
        - user-controller
      operationId: validateOTP
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/v1/token/refresh:
    post:
      tags:
        - user-controller
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTokenResponse"
  /api/v1/sendotp:
    post:
      tags:
        - user-controller
      operationId: sendOTP
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OTPRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /api/v1/resetpassword:
    post:
      tags:
        - user-controller
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /api/v1/registration:
    post:
      tags:
        - user-controller
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
  /api/v1/login:
    post:
      tags:
        - user-controller
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseTokenResponse"
  /api/v1/forgotpassword:
    post:
      tags:
        - user-controller
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseString"
components:
  schemas:
    OTPRequest:
      type: object
      properties:
        email:
          type: string
        otp:
          type: string
        type:
          type: string
          enum:
            - REGISTRATION
            - LOGIN
            - FORGOT_PASSWORD
    ApiResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
    ApiResponseTokenResponse:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: "#/components/schemas/TokenResponse"
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    ApiResponseString:
      type: object
      properties:
        status:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: string
    ResetPasswordRequest:
      type: object
      properties:
        email:
          type: string
        newPassword:
          type: string
    RegistrationRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    EmailRequest:
      type: object
      properties:
        email:
          type: string